add: rootスクリーンの切り出し
